Add files via upload
This notebook leverages Python-based data analytics to explore the relationship between national happiness, as reported in the World Happiness Report, and music preferences, as reflected in Spotify’s top-50 playlists for each country. The analysis, spanning from October 2023 to January 2024, seeks to uncover patterns and correlations that may shed light on the influence of music on a nation’s collective well-being.